refactor: requests 검색 API 코드 개선
- repository에 있던 기존 QueryBuilder 코드와 비즈니스 로직들을 service로 이전하였습니다.
